requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
docx2pdf==0.1.8
fastapi==0.135.1
ffmpeg-python==0.2.0
fire==0.7.1
fonttools==4.61.1
future==1.0.0
gunicorn==25.1.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pdf2docx==0.5.11
pdf2image==1.17.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyMuPDF==1.27.1
pypdfium2==5.6.0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311; platform_system == "Windows"
requests==2.32.5
starlette==0.52.1
termcolor==3.3.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
